requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
coverage==7.10.7
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mctools==1.3.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2==2.9.10
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-flask==1.3.0
python-dotenv==1.1.1
SQLAlchemy==2.0.35
typing_extensions==4.14.1
waitress==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
